feat(flutter_bloc)!: merge flutter_cubit into flutter_bloc
