+++ .github/workflows/bloc_test.yaml
@@ -31,6 +31,8 @@ jobs:
         run: pub run test_coverage
       - name: Check Code Coverage
         uses: ChicagoFlutter/lcov-cop@v1.0.0
+        with:
+          path: packages/bloc_test/coverage/lcov.info
       - name: Upload coverage to Codecov
         uses: codecov/codecov-action@v1
         with:
