+++ .github/workflows/flutter_bloc.yaml
@@ -23,13 +23,15 @@ jobs:
       - name: Install Dependencies
         run: flutter packages get
       - name: Format
-        run: flutter format --set-exit-if-changed lib test
+        run: flutter format --set-exit-if-changed lib test example
       - name: Analyze
-        run: flutter analyze lib test
+        run: flutter analyze lib test example
       - name: Run tests
         run: flutter test --no-pub --coverage --test-randomize-ordering-seed random
       - name: Check Code Coverage
         uses: ChicagoFlutter/lcov-cop@v1.0.0
+        with:
+          path: packages/flutter_bloc/coverage/lcov.info
       - name: Upload coverage to Codecov
         uses: codecov/codecov-action@v1
         with:
