--- .github/workflows/flutter_cubit.yaml
@@ -1,38 +1,0 @@
-name: flutter_cubit
-
-on:
-  push:
-    branches:
-      - master
-    paths:
-      - ".github/workflows/flutter_cubit.yaml"
-      - "packages/flutter_cubit/**"
-
-  pull_request:
-    branches:
-      - master
-    paths:
-      - ".github/workflows/flutter_cubit.yaml"
-      - "packages/flutter_cubit/**"
-
-jobs:
-  build:
-    defaults:
-      run:
-        working-directory: packages/flutter_cubit
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-      - uses: subosito/flutter-action@v1.3.0
-      - name: Install Dependencies
-        run: flutter packages get
-      - name: Format
-        run: flutter format --set-exit-if-changed lib test
-      - name: Analyze
-        run: flutter analyze lib test
-      - name: Run tests
-        run: flutter test --no-pub --coverage --test-randomize-ordering-seed random
-      - name: Check Code Coverage
-        uses: ChicagoFlutter/lcov-cop@v1.0.0
-        with:
-          path: packages/flutter_cubit/coverage/lcov.info
