+++ packages/flutter_bloc/lib/flutter_bloc.dart
@@ -4,8 +4,8 @@ export 'package:bloc/bloc.dart';
 
 export './src/bloc_builder.dart';
 export './src/bloc_consumer.dart';
-export './src/bloc_listener.dart';
-export './src/bloc_provider.dart';
+export './src/bloc_listener.dart' hide BlocListenerSingleChildWidget;
+export './src/bloc_provider.dart' hide BlocProviderSingleChildWidget;
 export './src/multi_bloc_listener.dart';
 export './src/multi_bloc_provider.dart';
 export './src/multi_repository_provider.dart';
