+++ packages/flutter_bloc/lib/src/repository_provider.dart
@@ -8,12 +8,12 @@ mixin RepositoryProviderSingleChildWidget on SingleChildWidget {}
 
 /// {@template repository_provider}
 /// Takes a `ValueBuilder` that is responsible for creating the repository and
-/// a [child] which will have access to the repository via
+/// a `child` which will have access to the repository via
 /// `RepositoryProvider.of(context)`.
 /// It is used as a dependency injection (DI) widget so that a single instance
 /// of a repository can be provided to multiple widgets within a subtree.
 ///
-/// Lazily creates the provided repository unless [lazy] is set to `false`.
+/// Lazily creates the provided repository unless `lazy` is set to `false`.
 ///
 /// ```dart
 /// RepositoryProvider(
@@ -41,7 +41,7 @@ class RepositoryProvider<T> extends Provider<T>
   /// Takes a repository and a [child] which will have access to the repository.
   /// A new repository should not be created in `RepositoryProvider.value`.
   /// Repositories should always be created using the default constructor
-  /// within the [builder].
+  /// within the `builder`.
   RepositoryProvider.value({
     Key key,
     @required T value,
@@ -59,14 +59,14 @@ class RepositoryProvider<T> extends Provider<T>
       return Provider.of<T>(context, listen: false);
     } on ProviderNotFoundException catch (_) {
       throw FlutterError(
-        """
+        '''
         RepositoryProvider.of() called with a context that does not contain a repository of type $T.
         No ancestor could be found starting from the context that was passed to RepositoryProvider.of<$T>().
 
         This can happen if the context you used comes from a widget above the RepositoryProvider.
 
         The context used was: $context
-        """,
+        ''',
       );
     }
   }
