+++ packages/flutter_bloc/test/bloc_builder_test.dart
@@ -1,64 +1,58 @@
-import 'dart:async';
-
+import 'package:bloc/bloc.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 class MyThemeApp extends StatefulWidget {
-  final Bloc<ThemeEvent, ThemeData> _themeBloc;
-  final Function _onBuild;
-
   MyThemeApp({
     Key key,
-    @required Bloc<ThemeEvent, ThemeData> themeBloc,
+    @required Cubit<ThemeData> themeCubit,
     @required Function onBuild,
-  })  : _themeBloc = themeBloc,
+  })  : _themeCubit = themeCubit,
         _onBuild = onBuild,
         super(key: key);
 
+  final Cubit<ThemeData> _themeCubit;
+  final Function _onBuild;
+
   @override
   State<MyThemeApp> createState() => MyThemeAppState(
-        themeBloc: _themeBloc,
+        themeCubit: _themeCubit,
         onBuild: _onBuild,
       );
 }
 
 class MyThemeAppState extends State<MyThemeApp> {
-  Bloc<ThemeEvent, ThemeData> _themeBloc;
-  final Function _onBuild;
-
   MyThemeAppState({
-    Key key,
-    @required Bloc<ThemeEvent, ThemeData> themeBloc,
+    @required Cubit<ThemeData> themeCubit,
     @required Function onBuild,
-  })  : _themeBloc = themeBloc,
+  })  : _themeCubit = themeCubit,
         _onBuild = onBuild;
+
+  Cubit<ThemeData> _themeCubit;
+  final Function _onBuild;
 
   @override
   Widget build(BuildContext context) {
-    return BlocBuilder(
-      bloc: _themeBloc,
+    return BlocBuilder<Cubit<ThemeData>, ThemeData>(
+      cubit: _themeCubit,
       builder: ((context, theme) {
         _onBuild();
         return MaterialApp(
-          key: Key('material_app'),
+          key: const Key('material_app'),
           theme: theme,
           home: Column(
             children: [
               RaisedButton(
-                key: Key('raised_button_1'),
+                key: const Key('raised_button_1'),
                 onPressed: () {
-                  setState(() {
-                    _themeBloc = DarkThemeBloc();
-                  });
+                  setState(() => _themeCubit = DarkThemeCubit());
                 },
               ),
               RaisedButton(
-                key: Key('raised_button_2'),
+                key: const Key('raised_button_2'),
                 onPressed: () {
-                  setState(() {
-                    _themeBloc = _themeBloc;
-                  });
+                  setState(() => _themeCubit = _themeCubit);
                 },
               ),
             ],
@@ -69,28 +63,18 @@ class MyThemeAppState extends State<MyThemeApp> {
   }
 }
 
-abstract class ThemeEvent {}
+class ThemeCubit extends Cubit<ThemeData> {
+  ThemeCubit() : super(ThemeData.light());
 
-class SetDarkTheme extends ThemeEvent {}
-
-class SetLightTheme extends ThemeEvent {}
-
-class ThemeBloc extends Bloc<ThemeEvent, ThemeData> {
-  ThemeBloc() : super(ThemeData.light());
-
-  @override
-  Stream<ThemeData> mapEventToState(ThemeEvent event) async* {
-    yield event is SetDarkTheme ? ThemeData.dark() : ThemeData.light();
-  }
+  void setDarkTheme() => emit(ThemeData.dark());
+  void setLightTheme() => emit(ThemeData.light());
 }
 
-class DarkThemeBloc extends Bloc<ThemeEvent, ThemeData> {
-  DarkThemeBloc() : super(ThemeData.dark());
+class DarkThemeCubit extends Cubit<ThemeData> {
+  DarkThemeCubit() : super(ThemeData.dark());
 
-  @override
-  Stream<ThemeData> mapEventToState(ThemeEvent event) async* {
-    yield event is SetDarkTheme ? ThemeData.dark() : ThemeData.light();
-  }
+  void setDarkTheme() => emit(ThemeData.dark());
+  void setLightTheme() => emit(ThemeData.light());
 }
 
 class MyCounterApp extends StatefulWidget {
@@ -99,35 +83,39 @@ class MyCounterApp extends StatefulWidget {
 }
 
 class MyCounterAppState extends State<MyCounterApp> {
-  final CounterBloc _bloc = CounterBloc();
+  final CounterCubit _cubit = CounterCubit();
 
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
       home: Scaffold(
-        key: Key('myCounterApp'),
+        key: const Key('myCounterApp'),
         body: Column(
           children: <Widget>[
-            BlocBuilder<CounterBloc, int>(
-              bloc: _bloc,
+            BlocBuilder<CounterCubit, int>(
+              cubit: _cubit,
               buildWhen: (previousState, state) {
                 return (previousState + state) % 3 == 0;
               },
               builder: (context, count) {
-                return Text('$count', key: Key('myCounterAppTextCondition'));
+                return Text(
+                  '$count',
+                  key: const Key('myCounterAppTextCondition'),
+                );
               },
             ),
-            BlocBuilder<CounterBloc, int>(
-              bloc: _bloc,
+            BlocBuilder<CounterCubit, int>(
+              cubit: _cubit,
               builder: (context, count) {
-                return Text('$count', key: Key('myCounterAppText'));
+                return Text(
+                  '$count',
+                  key: const Key('myCounterAppText'),
+                );
               },
             ),
             RaisedButton(
-              key: Key('myCounterAppIncrementButton'),
-              onPressed: () {
-                _bloc.add(CounterEvent.increment);
-              },
+              key: const Key('myCounterAppIncrementButton'),
+              onPressed: _cubit.increment,
             )
           ],
         ),
@@ -136,32 +124,21 @@ class MyCounterAppState extends State<MyCounterApp> {
   }
 }
 
-enum CounterEvent { increment, decrement }
+class CounterCubit extends Cubit<int> {
+  CounterCubit() : super(0);
 
-class CounterBloc extends Bloc<CounterEvent, int> {
-  CounterBloc() : super(0);
-
-  @override
-  Stream<int> mapEventToState(CounterEvent event) async* {
-    switch (event) {
-      case CounterEvent.decrement:
-        yield state - 1;
-        break;
-      case CounterEvent.increment:
-        yield state + 1;
-        break;
-    }
-  }
+  void increment() => emit(state + 1);
+  void decrement() => emit(state - 1);
 }
 
 void main() {
   group('BlocBuilder', () {
-    testWidgets('throws if initialized with null bloc and builder',
+    testWidgets('throws if initialized with null cubit and builder',
         (tester) async {
       try {
         await tester.pumpWidget(
-          BlocBuilder<ThemeBloc, ThemeData>(
-            bloc: null,
+          BlocBuilder<ThemeCubit, ThemeData>(
+            cubit: null,
             builder: null,
           ),
         );
@@ -173,8 +150,8 @@ void main() {
     testWidgets('throws if initialized with null builder', (tester) async {
       try {
         await tester.pumpWidget(
-          BlocBuilder<ThemeBloc, ThemeData>(
-            bloc: ThemeBloc(),
+          BlocBuilder<ThemeCubit, ThemeData>(
+            cubit: ThemeCubit(),
             builder: null,
           ),
         );
@@ -184,203 +161,178 @@ void main() {
     });
 
     testWidgets('passes initial state to widget', (tester) async {
-      final _themeBloc = ThemeBloc();
+      final themeCubit = ThemeCubit();
       var numBuilds = 0;
       await tester.pumpWidget(
-        MyThemeApp(
-          themeBloc: _themeBloc,
-          onBuild: () {
-            numBuilds++;
-          },
-        ),
+        MyThemeApp(themeCubit: themeCubit, onBuild: () => numBuilds++),
       );
 
-      final _materialApp = find
-          .byKey(Key('material_app'))
-          .evaluate()
-          .first
-          .widget as MaterialApp;
-      expect(_materialApp.theme, ThemeData.light());
+      final materialApp = tester.widget<MaterialApp>(
+        find.byKey(const Key('material_app')),
+      );
+
+      expect(materialApp.theme, ThemeData.light());
       expect(numBuilds, 1);
     });
 
     testWidgets('receives events and sends state updates to widget',
         (tester) async {
-      final _themeBloc = ThemeBloc();
+      final themeCubit = ThemeCubit();
       var numBuilds = 0;
       await tester.pumpWidget(
-        MyThemeApp(
-          themeBloc: _themeBloc,
-          onBuild: () {
-            numBuilds++;
-          },
-        ),
+        MyThemeApp(themeCubit: themeCubit, onBuild: () => numBuilds++),
       );
 
-      _themeBloc.add(SetDarkTheme());
+      themeCubit.setDarkTheme();
 
       await tester.pumpAndSettle();
 
-      final _materialApp = find
-          .byKey(Key('material_app'))
-          .evaluate()
-          .first
-          .widget as MaterialApp;
+      final materialApp = tester.widget<MaterialApp>(
+        find.byKey(const Key('material_app')),
+      );
 
-      expect(_materialApp.theme, ThemeData.dark());
+      expect(materialApp.theme, ThemeData.dark());
       expect(numBuilds, 2);
     });
 
-    testWidgets('infers the bloc from the context if the bloc is not provided',
+    testWidgets(
+        'infers the cubit from the context if the cubit is not provided',
         (tester) async {
-      final themeBloc = ThemeBloc();
+      final themeCubit = ThemeCubit();
       var numBuilds = 0;
       await tester.pumpWidget(
         BlocProvider.value(
-          value: themeBloc,
-          child: BlocBuilder<ThemeBloc, ThemeData>(
-            builder: (
-              context,
-              theme,
-            ) {
+          value: themeCubit,
+          child: BlocBuilder<ThemeCubit, ThemeData>(
+            builder: (context, theme) {
               numBuilds++;
               return MaterialApp(
-                key: Key('material_app'),
+                key: const Key('material_app'),
                 theme: theme,
-                home: Container(),
+                home: const SizedBox(),
               );
             },
           ),
         ),
       );
 
-      themeBloc.add(SetDarkTheme());
+      themeCubit.setDarkTheme();
 
-      await tester.pump();
+      await tester.pumpAndSettle();
 
-      var _materialApp = find.byKey(Key('material_app')).evaluate().first.widget
-          as MaterialApp;
+      var materialApp = tester.widget<MaterialApp>(
+        find.byKey(const Key('material_app')),
+      );
 
-      expect(_materialApp.theme, ThemeData.dark());
+      expect(materialApp.theme, ThemeData.dark());
       expect(numBuilds, 2);
 
-      themeBloc.add(SetLightTheme());
+      themeCubit.setLightTheme();
 
-      await tester.pump();
+      await tester.pumpAndSettle();
 
-      _materialApp = find.byKey(Key('material_app')).evaluate().first.widget
-          as MaterialApp;
+      materialApp = tester.widget<MaterialApp>(
+        find.byKey(const Key('material_app')),
+      );
 
-      expect(_materialApp.theme, ThemeData.light());
+      expect(materialApp.theme, ThemeData.light());
       expect(numBuilds, 3);
     });
 
     testWidgets(
-        'updates when the bloc is changed at runtime to a different bloc and'
-        'unsubscribes from old bloc', (tester) async {
-      final _themeBloc = ThemeBloc();
+        'updates when the cubit is changed at runtime to a different cubit and'
+        'unsubscribes from old cubit', (tester) async {
+      final themeCubit = ThemeCubit();
       var numBuilds = 0;
       await tester.pumpWidget(
-        MyThemeApp(
-          themeBloc: _themeBloc,
-          onBuild: () {
-            numBuilds++;
-          },
-        ),
+        MyThemeApp(themeCubit: themeCubit, onBuild: () => numBuilds++),
       );
 
       await tester.pumpAndSettle();
 
-      var _materialApp = find.byKey(Key('material_app')).evaluate().first.widget
-          as MaterialApp;
+      var materialApp = tester.widget<MaterialApp>(
+        find.byKey(const Key('material_app')),
+      );
 
-      expect(_materialApp.theme, ThemeData.light());
+      expect(materialApp.theme, ThemeData.light());
       expect(numBuilds, 1);
 
-      await tester.tap(find.byKey(Key('raised_button_1')));
+      await tester.tap(find.byKey(const Key('raised_button_1')));
       await tester.pumpAndSettle();
 
-      _materialApp = find.byKey(Key('material_app')).evaluate().first.widget
-          as MaterialApp;
+      materialApp = tester.widget<MaterialApp>(
+        find.byKey(const Key('material_app')),
+      );
 
-      expect(_materialApp.theme, ThemeData.dark());
+      expect(materialApp.theme, ThemeData.dark());
       expect(numBuilds, 2);
 
-      _themeBloc.add(SetLightTheme());
+      themeCubit.setLightTheme();
       await tester.pumpAndSettle();
 
-      _materialApp = find.byKey(Key('material_app')).evaluate().first.widget
-          as MaterialApp;
+      materialApp = tester.widget<MaterialApp>(
+        find.byKey(const Key('material_app')),
+      );
 
-      expect(_materialApp.theme, ThemeData.dark());
+      expect(materialApp.theme, ThemeData.dark());
       expect(numBuilds, 2);
     });
 
     testWidgets(
-        'does not update when the bloc is changed at runtime to same bloc '
-        'and stays subscribed to current bloc', (tester) async {
-      final _themeBloc = DarkThemeBloc();
+        'does not update when the cubit is changed at runtime to same cubit '
+        'and stays subscribed to current cubit', (tester) async {
+      final themeCubit = DarkThemeCubit();
       var numBuilds = 0;
       await tester.pumpWidget(
-        MyThemeApp(
-          themeBloc: _themeBloc,
-          onBuild: () {
-            numBuilds++;
-          },
-        ),
+        MyThemeApp(themeCubit: themeCubit, onBuild: () => numBuilds++),
       );
 
       await tester.pumpAndSettle();
 
-      var _materialApp = find.byKey(Key('material_app')).evaluate().first.widget
-          as MaterialApp;
+      var materialApp = tester.widget<MaterialApp>(
+        find.byKey(const Key('material_app')),
+      );
 
-      expect(_materialApp.theme, ThemeData.dark());
+      expect(materialApp.theme, ThemeData.dark());
       expect(numBuilds, 1);
 
-      await tester.tap(find.byKey(Key('raised_button_2')));
+      await tester.tap(find.byKey(const Key('raised_button_2')));
       await tester.pumpAndSettle();
 
-      _materialApp = find.byKey(Key('material_app')).evaluate().first.widget
-          as MaterialApp;
+      materialApp = tester.widget<MaterialApp>(
+        find.byKey(const Key('material_app')),
+      );
 
-      expect(_materialApp.theme, ThemeData.dark());
+      expect(materialApp.theme, ThemeData.dark());
       expect(numBuilds, 2);
 
-      _themeBloc.add(SetLightTheme());
+      themeCubit.setLightTheme();
       await tester.pumpAndSettle();
 
-      _materialApp = find.byKey(Key('material_app')).evaluate().first.widget
-          as MaterialApp;
+      materialApp = tester.widget<MaterialApp>(
+        find.byKey(const Key('material_app')),
+      );
 
-      expect(_materialApp.theme, ThemeData.light());
+      expect(materialApp.theme, ThemeData.light());
       expect(numBuilds, 3);
     });
 
     testWidgets('shows latest state instead of initial state', (tester) async {
-      final _themeBloc = ThemeBloc();
-      _themeBloc.add(SetDarkTheme());
+      final themeCubit = ThemeCubit()..setDarkTheme();
       await tester.pumpAndSettle();
 
       var numBuilds = 0;
       await tester.pumpWidget(
-        MyThemeApp(
-          themeBloc: _themeBloc,
-          onBuild: () {
-            numBuilds++;
-          },
-        ),
+        MyThemeApp(themeCubit: themeCubit, onBuild: () => numBuilds++),
       );
 
       await tester.pumpAndSettle();
 
-      final _materialApp = find
-          .byKey(Key('material_app'))
-          .evaluate()
-          .first
-          .widget as MaterialApp;
+      final materialApp = tester.widget<MaterialApp>(
+        find.byKey(const Key('material_app')),
+      );
 
-      expect(_materialApp.theme, ThemeData.dark());
+      expect(materialApp.theme, ThemeData.dark());
       expect(numBuilds, 1);
     });
 
@@ -389,51 +341,51 @@ void main() {
       await tester.pumpWidget(MyCounterApp());
       await tester.pumpAndSettle();
 
-      expect(find.byKey(Key('myCounterApp')), findsOneWidget);
+      expect(find.byKey(const Key('myCounterApp')), findsOneWidget);
 
       final incrementButtonFinder =
-          find.byKey(Key('myCounterAppIncrementButton'));
+          find.byKey(const Key('myCounterAppIncrementButton'));
       expect(incrementButtonFinder, findsOneWidget);
 
       final counterText1 =
-          tester.widget(find.byKey(Key('myCounterAppText'))) as Text;
+          tester.widget<Text>(find.byKey(const Key('myCounterAppText')));
       expect(counterText1.data, '0');
 
-      final conditionalCounterText1 =
-          tester.widget(find.byKey(Key('myCounterAppTextCondition'))) as Text;
+      final conditionalCounterText1 = tester
+          .widget<Text>(find.byKey(const Key('myCounterAppTextCondition')));
       expect(conditionalCounterText1.data, '0');
 
       await tester.tap(incrementButtonFinder);
       await tester.pumpAndSettle();
 
       final counterText2 =
-          tester.widget(find.byKey(Key('myCounterAppText'))) as Text;
+          tester.widget<Text>(find.byKey(const Key('myCounterAppText')));
       expect(counterText2.data, '1');
 
-      final conditionalCounterText2 =
-          tester.widget(find.byKey(Key('myCounterAppTextCondition'))) as Text;
+      final conditionalCounterText2 = tester
+          .widget<Text>(find.byKey(const Key('myCounterAppTextCondition')));
       expect(conditionalCounterText2.data, '0');
 
       await tester.tap(incrementButtonFinder);
       await tester.pumpAndSettle();
 
       final counterText3 =
-          tester.widget(find.byKey(Key('myCounterAppText'))) as Text;
+          tester.widget<Text>(find.byKey(const Key('myCounterAppText')));
       expect(counterText3.data, '2');
 
-      final conditionalCounterText3 =
-          tester.widget(find.byKey(Key('myCounterAppTextCondition'))) as Text;
+      final conditionalCounterText3 = tester
+          .widget<Text>(find.byKey(const Key('myCounterAppTextCondition')));
       expect(conditionalCounterText3.data, '2');
 
       await tester.tap(incrementButtonFinder);
       await tester.pumpAndSettle();
 
       final counterText4 =
-          tester.widget(find.byKey(Key('myCounterAppText'))) as Text;
+          tester.widget<Text>(find.byKey(const Key('myCounterAppText')));
       expect(counterText4.data, '3');
 
-      final conditionalCounterText4 =
-          tester.widget(find.byKey(Key('myCounterAppTextCondition'))) as Text;
+      final conditionalCounterText4 = tester
+          .widget<Text>(find.byKey(const Key('myCounterAppTextCondition')));
       expect(conditionalCounterText4.data, '2');
     });
   });
