+++ packages/flutter_bloc/test/bloc_consumer_test.dart
@@ -1,24 +1,13 @@
+import 'package:bloc/bloc.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-enum CounterEvent { increment, decrement }
+class CounterCubit extends Cubit<int> {
+  CounterCubit() : super(0);
 
-class CounterBloc extends Bloc<CounterEvent, int> {
-  CounterBloc() : super(0);
-
-  @override
-  Stream<int> mapEventToState(CounterEvent event) async* {
-    switch (event) {
-      case CounterEvent.decrement:
-        yield state - 1;
-        break;
-      case CounterEvent.increment:
-        yield state + 1;
-        break;
-    }
-  }
+  void increment() => emit(state + 1);
 }
 
 void main() {
@@ -26,7 +15,7 @@ void main() {
     testWidgets('throws AssertionError if builder is null', (tester) async {
       try {
         await tester.pumpWidget(
-          BlocConsumer(
+          BlocConsumer<CounterCubit, int>(
             builder: null,
             listener: (_, __) {},
           ),
@@ -39,8 +28,8 @@ void main() {
     testWidgets('throws AssertionError if listener is null', (tester) async {
       try {
         await tester.pumpWidget(
-          BlocConsumer(
-            builder: (_, __) => Container(),
+          BlocConsumer<CounterCubit, int>(
+            builder: (_, __) => const SizedBox(),
             listener: null,
           ),
         );
@@ -52,13 +41,13 @@ void main() {
     testWidgets(
         'accesses the bloc directly and passes initial state to builder and '
         'nothing to listener', (tester) async {
-      final counterBloc = CounterBloc();
+      final counterCubit = CounterCubit();
       final listenerStates = <int>[];
       await tester.pumpWidget(
         MaterialApp(
           home: Scaffold(
-            body: BlocConsumer(
-              bloc: counterBloc,
+            body: BlocConsumer<CounterCubit, int>(
+              cubit: counterCubit,
               builder: (context, state) {
                 return Text('State: $state');
               },
@@ -76,13 +65,13 @@ void main() {
     testWidgets(
         'accesses the bloc directly '
         'and passes multiple states to builder and listener', (tester) async {
-      final counterBloc = CounterBloc();
+      final counterCubit = CounterCubit();
       final listenerStates = <int>[];
       await tester.pumpWidget(
         MaterialApp(
           home: Scaffold(
-            body: BlocConsumer(
-              bloc: counterBloc,
+            body: BlocConsumer<CounterCubit, int>(
+              cubit: counterCubit,
               builder: (context, state) {
                 return Text('State: $state');
               },
@@ -95,7 +84,7 @@ void main() {
       );
       expect(find.text('State: 0'), findsOneWidget);
       expect(listenerStates, isEmpty);
-      counterBloc.add(CounterEvent.increment);
+      counterCubit.increment();
       await tester.pump();
       expect(find.text('State: 1'), findsOneWidget);
       expect(listenerStates, [1]);
@@ -104,15 +93,15 @@ void main() {
     testWidgets(
         'accesses the bloc via context and passes initial state to builder',
         (tester) async {
-      final counterBloc = CounterBloc();
+      final counterCubit = CounterCubit();
       final listenerStates = <int>[];
       await tester.pumpWidget(
-        BlocProvider<CounterBloc>.value(
-          value: counterBloc,
+        BlocProvider<CounterCubit>.value(
+          value: counterCubit,
           child: MaterialApp(
             home: Scaffold(
-              body: BlocConsumer(
-                bloc: counterBloc,
+              body: BlocConsumer<CounterCubit, int>(
+                cubit: counterCubit,
                 builder: (context, state) {
                   return Text('State: $state');
                 },
@@ -131,13 +120,13 @@ void main() {
     testWidgets(
         'accesses the bloc via context and passes multiple states to builder',
         (tester) async {
-      final counterBloc = CounterBloc();
+      final counterCubit = CounterCubit();
       final listenerStates = <int>[];
       await tester.pumpWidget(
         MaterialApp(
           home: Scaffold(
-            body: BlocConsumer(
-              bloc: counterBloc,
+            body: BlocConsumer<CounterCubit, int>(
+              cubit: counterCubit,
               builder: (context, state) {
                 return Text('State: $state');
               },
@@ -150,7 +139,7 @@ void main() {
       );
       expect(find.text('State: 0'), findsOneWidget);
       expect(listenerStates, isEmpty);
-      counterBloc.add(CounterEvent.increment);
+      counterCubit.increment();
       await tester.pump();
       expect(find.text('State: 1'), findsOneWidget);
       expect(listenerStates, [1]);
@@ -158,14 +147,14 @@ void main() {
 
     testWidgets('does not trigger rebuilds when buildWhen evaluates to false',
         (tester) async {
-      final counterBloc = CounterBloc();
+      final counterCubit = CounterCubit();
       final listenerStates = <int>[];
       final builderStates = <int>[];
       await tester.pumpWidget(
         MaterialApp(
           home: Scaffold(
-            body: BlocConsumer(
-              bloc: counterBloc,
+            body: BlocConsumer<CounterCubit, int>(
+              cubit: counterCubit,
               buildWhen: (previous, current) => (previous + current) % 3 == 0,
               builder: (context, state) {
                 builderStates.add(state);
@@ -182,14 +171,14 @@ void main() {
       expect(builderStates, [0]);
       expect(listenerStates, isEmpty);
 
-      counterBloc.add(CounterEvent.increment);
+      counterCubit.increment();
       await tester.pump();
 
       expect(find.text('State: 0'), findsOneWidget);
       expect(builderStates, [0]);
       expect(listenerStates, [1]);
 
-      counterBloc.add(CounterEvent.increment);
+      counterCubit.increment();
       await tester.pumpAndSettle();
 
       expect(find.text('State: 2'), findsOneWidget);
@@ -199,14 +188,14 @@ void main() {
 
     testWidgets('does not trigger listen when listenWhen evaluates to false',
         (tester) async {
-      final counterBloc = CounterBloc();
+      final counterCubit = CounterCubit();
       final listenerStates = <int>[];
       final builderStates = <int>[];
       await tester.pumpWidget(
         MaterialApp(
           home: Scaffold(
-            body: BlocConsumer(
-              bloc: counterBloc,
+            body: BlocConsumer<CounterCubit, int>(
+              cubit: counterCubit,
               builder: (context, state) {
                 builderStates.add(state);
                 return Text('State: $state');
@@ -223,14 +212,14 @@ void main() {
       expect(builderStates, [0]);
       expect(listenerStates, isEmpty);
 
-      counterBloc.add(CounterEvent.increment);
+      counterCubit.increment();
       await tester.pump();
 
       expect(find.text('State: 1'), findsOneWidget);
       expect(builderStates, [0, 1]);
       expect(listenerStates, isEmpty);
 
-      counterBloc.add(CounterEvent.increment);
+      counterCubit.increment();
       await tester.pumpAndSettle();
 
       expect(find.text('State: 2'), findsOneWidget);
