+++ packages/flutter_bloc/test/repository_provider_test.dart
@@ -3,45 +3,40 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 class MyApp extends StatelessWidget {
-  final Repository _repository;
-  final Widget _child;
-  final bool useValueProvider;
-
   const MyApp({
     Key key,
-    @required Repository repository,
-    @required Widget child,
+    @required this.repository,
+    @required this.child,
     this.useValueProvider,
-  })  : _repository = repository,
-        _child = child,
-        super(key: key);
+  }) : super(key: key);
+
+  final Repository repository;
+  final Widget child;
+  final bool useValueProvider;
 
   @override
   Widget build(BuildContext context) {
     if (useValueProvider == true) {
       return MaterialApp(
         home: RepositoryProvider<Repository>.value(
-          value: _repository,
-          child: _child,
+          value: repository,
+          child: child,
         ),
       );
     }
     return MaterialApp(
       home: RepositoryProvider<Repository>(
-        create: (context) => _repository,
-        child: _child,
+        create: (_) => repository,
+        child: child,
       ),
     );
   }
 }
 
 class MyStatefulApp extends StatefulWidget {
-  final Widget child;
+  const MyStatefulApp({Key key, @required this.child}) : super(key: key);
 
-  const MyStatefulApp({
-    Key key,
-    @required this.child,
-  }) : super(key: key);
+  final Widget child;
 
   @override
   _MyStatefulAppState createState() => _MyStatefulAppState();
@@ -52,7 +47,7 @@ class _MyStatefulAppState extends State<MyStatefulApp> {
 
   @override
   void initState() {
-    _repository = Repository(0);
+    _repository = const Repository(0);
     super.initState();
   }
 
@@ -60,19 +55,15 @@ class _MyStatefulAppState extends State<MyStatefulApp> {
   Widget build(BuildContext context) {
     return MaterialApp(
       home: RepositoryProvider<Repository>(
-        create: (context) => _repository,
+        create: (_) => _repository,
         child: Scaffold(
           appBar: AppBar(
-            title: Text('Counter'),
             actions: <Widget>[
               IconButton(
-                key: Key('iconButtonKey'),
-                icon: Icon(Icons.edit),
-                tooltip: 'Change State',
+                key: const Key('iconButtonKey'),
+                icon: const Icon(Icons.edit),
                 onPressed: () {
-                  setState(() {
-                    _repository = Repository(0);
-                  });
+                  setState(() => _repository = const Repository(0));
                 },
               )
             ],
@@ -84,27 +75,15 @@ class _MyStatefulAppState extends State<MyStatefulApp> {
   }
 }
 
-class MyAppNoProvider extends StatelessWidget {
-  final Widget _child;
-
-  const MyAppNoProvider({
-    Key key,
-    @required Widget child,
-  })  : _child = child,
-        super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      home: _child,
-    );
-  }
+class MyAppNoProvider extends MaterialApp {
+  const MyAppNoProvider({Key key, @required Widget child})
+      : super(key: key, home: child);
 }
 
 class CounterPage extends StatelessWidget {
-  final Function onBuild;
-
   const CounterPage({Key key, this.onBuild}) : super(key: key);
+
+  final VoidCallback onBuild;
 
   @override
   Widget build(BuildContext context) {
@@ -113,14 +92,7 @@ class CounterPage extends StatelessWidget {
     assert(repository != null);
 
     return Scaffold(
-      appBar: AppBar(title: Text('Value')),
-      body: Center(
-        child: Text(
-          '${repository.data}',
-          key: Key('value_data'),
-          style: TextStyle(fontSize: 24.0),
-        ),
-      ),
+      body: Text('${repository.data}', key: const Key('value_data')),
     );
   }
 }
@@ -130,10 +102,10 @@ class RoutePage extends StatelessWidget {
   Widget build(BuildContext context) {
     return Scaffold(
       body: RaisedButton(
-        key: Key('route_button'),
+        key: const Key('route_button'),
         onPressed: () {
           Navigator.of(context).pushReplacement(
-            MaterialPageRoute<Widget>(builder: (context) => Container()),
+            MaterialPageRoute<Widget>(builder: (context) => const SizedBox()),
           );
         },
       ),
@@ -142,26 +114,24 @@ class RoutePage extends StatelessWidget {
 }
 
 class Repository {
-  final int data;
+  const Repository(this.data);
 
-  Repository(this.data);
+  final int data;
 }
 
 void main() {
   group('RepositoryProvider', () {
     testWidgets('throws if initialized with no repository', (tester) async {
-      await tester.pumpWidget(MyApp(
-        repository: null,
-        child: CounterPage(),
-      ));
+      await tester.pumpWidget(
+        const MyApp(repository: null, child: CounterPage()),
+      );
       expect(tester.takeException(), isInstanceOf<AssertionError>());
     });
 
     testWidgets('throws if initialized with no child', (tester) async {
-      await tester.pumpWidget(MyApp(
-        repository: Repository(0),
-        child: null,
-      ));
+      await tester.pumpWidget(
+        const MyApp(repository: Repository(0), child: null),
+      );
       expect(tester.takeException(), isInstanceOf<AssertionError>());
     });
 
@@ -171,9 +141,9 @@ void main() {
         RepositoryProvider(
           create: (_) {
             createCalled = true;
-            return Repository(0);
+            return const Repository(0);
           },
-          child: Container(),
+          child: const SizedBox(),
         ),
       );
       expect(createCalled, isFalse);
@@ -185,24 +155,23 @@ void main() {
         RepositoryProvider(
           create: (_) {
             createCalled = true;
-            return Repository(0);
+            return const Repository(0);
           },
           lazy: false,
-          child: Container(),
+          child: const SizedBox(),
         ),
       );
       expect(createCalled, isTrue);
     });
 
     testWidgets('passes value to children via builder', (tester) async {
-      final repository = Repository(0);
-      final _child = CounterPage();
-      await tester.pumpWidget(MyApp(
-        repository: repository,
-        child: _child,
-      ));
+      const repository = Repository(0);
+      const child = CounterPage();
+      await tester.pumpWidget(
+        const MyApp(repository: repository, child: child),
+      );
 
-      final _counterFinder = find.byKey((Key('value_data')));
+      final _counterFinder = find.byKey((const Key('value_data')));
       expect(_counterFinder, findsOneWidget);
 
       final _counterText = _counterFinder.evaluate().first.widget as Text;
@@ -210,15 +179,15 @@ void main() {
     });
 
     testWidgets('passes value to children via value', (tester) async {
-      final repository = Repository(0);
-      final _child = CounterPage();
-      await tester.pumpWidget(MyApp(
+      const repository = Repository(0);
+      const child = CounterPage();
+      await tester.pumpWidget(const MyApp(
         repository: repository,
-        child: _child,
+        child: child,
         useValueProvider: true,
       ));
 
-      final _counterFinder = find.byKey((Key('value_data')));
+      final _counterFinder = find.byKey((const Key('value_data')));
       expect(_counterFinder, findsOneWidget);
 
       final _counterText = _counterFinder.evaluate().first.widget as Text;
@@ -228,19 +197,17 @@ void main() {
     testWidgets(
         'should throw FlutterError if RepositoryProvider is not found in '
         'current context', (tester) async {
-      final Widget _child = CounterPage();
-      await tester.pumpWidget(MyAppNoProvider(
-        child: _child,
-      ));
+      const child = CounterPage();
+      await tester.pumpWidget(const MyAppNoProvider(child: child));
       final dynamic exception = tester.takeException();
-      final expectedMessage = """
+      final expectedMessage = '''
         RepositoryProvider.of() called with a context that does not contain a repository of type Repository.
         No ancestor could be found starting from the context that was passed to RepositoryProvider.of<Repository>().
 
         This can happen if the context you used comes from a widget above the RepositoryProvider.
 
         The context used was: CounterPage(dirty)
-""";
+''';
       expect(exception is FlutterError, true);
       expect(exception.message, expectedMessage);
     });
@@ -253,7 +220,7 @@ void main() {
         RepositoryProvider<Repository>(
           lazy: false,
           create: (_) => throw expectedException,
-          child: Container(),
+          child: const SizedBox(),
         ),
       );
       final dynamic exception = tester.takeException();
@@ -264,15 +231,9 @@ void main() {
         'should not rebuild widgets that inherited the value if the value is '
         'changed', (tester) async {
       var numBuilds = 0;
-      final Widget _child = CounterPage(
-        onBuild: () {
-          numBuilds++;
-        },
-      );
-      await tester.pumpWidget(MyStatefulApp(
-        child: _child,
-      ));
-      await tester.tap(find.byKey(Key('iconButtonKey')));
+      final child = CounterPage(onBuild: () => numBuilds++);
+      await tester.pumpWidget(MyStatefulApp(child: child));
+      await tester.tap(find.byKey(const Key('iconButtonKey')));
       await tester.pump();
       expect(numBuilds, 1);
     });
@@ -282,16 +243,14 @@ void main() {
         'via RepositoryProviderExtension', (tester) async {
       await tester.pumpWidget(
         RepositoryProvider(
-          create: (_) => Repository(0),
+          create: (_) => const Repository(0),
           child: MaterialApp(
             home: Scaffold(
-              appBar: AppBar(title: Text('Value')),
               body: Center(
                 child: Builder(
                   builder: (context) => Text(
                     '${context.repository<Repository>().data}',
-                    key: Key('value_data'),
-                    style: TextStyle(fontSize: 24.0),
+                    key: const Key('value_data'),
                   ),
                 ),
               ),
@@ -299,7 +258,7 @@ void main() {
           ),
         ),
       );
-      final _counterFinder = find.byKey((Key('value_data')));
+      final _counterFinder = find.byKey((const Key('value_data')));
       expect(_counterFinder, findsOneWidget);
 
       final _counterText = _counterFinder.evaluate().first.widget as Text;
