--- packages/flutter_cubit/CHANGELOG.md
@@ -1,72 +1,0 @@
-# 0.1.1
-
-- fix: upgrade to `cubit: ^0.1.2`
-- docs: minor documentation updates
-
-# 0.1.0
-
-- **BREAKING**: rename `condition` in `CubitBuilder` to `buildWhen`
-- **BREAKING**: rename `condition` in `CubitListener` to `listenWhen`
-- feat: upgrade to `cubit: ^0.1.0`
-- docs: inline documentation improvements
-
-# 0.0.12
-
-- feat: upgrade to `cubit: ^0.0.13`
-- docs: various documentation updates
-
-# 0.0.11
-
-- feat: upgrade to `cubit ^0.0.10`
-- docs: various documentation updates
-
-# 0.0.10
-
-- **BREAKING**: upgrade to `cubit: ^0.0.8`
-- docs: minor logo updates
-
-# 0.0.9
-
-- feat: upgrade to `cubit: ^0.0.7`
-- feat: make all widgets compatible with `CubitStream`
-
-# 0.0.8
-
-- **BREAKING**: upgrade to `cubit: ^0.0.4`
-
-# 0.0.7
-
-- fix: use `Cubit<Object>` instead of `Cubit<dynamic>` in `CubitProvider`
-
-# 0.0.6
-
-- fix: expose `CubitListenerSingleChildWidget`
-- fix: expose `CubitProviderSingleChildWidget`
-
-# 0.0.5
-
-- fix: hide `CubitListenerSingleChildWidget`
-- fix: hide `CubitProviderSingleChildWidget`
-- fix: `ProviderNotFoundException` handling
-- fix: do not leak `Provider` implementation
-
-# 0.0.4
-
-- feat: add `CubitConsumer`
-- feat: add `MultiCubitListener`
-- feat: add `MultiCubitProvider`
-
-# 0.0.3
-
-- **BREAKING**: upgrade to `cubit: ^0.0.3`
-- docs: minor documentation improvements
-
-# 0.0.2
-
-- **BREAKING**: upgrade to `cubit: ^0.0.2`
-- docs: minor documentation improvements
-- tests: 100% code coverage
-
-# 0.0.1
-
-- feat: initial release
