--- packages/flutter_cubit/example/.metadata
@@ -1,10 +1,0 @@
-# This file tracks properties of this Flutter project.
-# Used by Flutter tool to assess capabilities and perform upgrades etc.
-#
-# This file should be version controlled and should not be manually edited.
-
-version:
-  revision: 5f21edf8b66e31a39133177319414395cc5b5f48
-  channel: stable
-
-project_type: app
