--- packages/flutter_cubit/example/pubspec.yaml
@@ -1,16 +1,0 @@
-name: example
-description: A new Flutter project.
-version: 1.0.0+1
-
-environment:
-  sdk: ">=2.0.0 <3.0.0"
-
-dependencies:
-  flutter:
-    sdk: flutter
-  cubit: ^0.1.0
-  flutter_cubit:
-    path: ../
-
-flutter:
-  uses-material-design: true
