--- packages/flutter_cubit/lib/flutter_cubit.dart
@@ -1,8 +1,0 @@
-library flutter_cubit;
-
-export 'src/cubit_builder.dart';
-export 'src/cubit_consumer.dart';
-export 'src/cubit_listener.dart';
-export 'src/cubit_provider.dart';
-export 'src/multi_cubit_listener.dart';
-export 'src/multi_cubit_provider.dart';
